backend/requirements.txt
# Website
Django>=3.2,<4.0
wagtail
# REST framework
djangorestframework
dj-rest-auth
django-filter
django-cors-headers
drf_spectacular
drf-spectacular[sidecar]
djangorestframework-yaml
drf-writable-nested
# Wagtail plugins
wagtailmedia
# Dependencies
psycopg2
pyyaml
# Deployment
whitenoise
uwsgi
# Development
pylint
black
pylint-django
